test: move integration/functional/extreme suites to tests-extensive/ (off default path)
Extract Kodi mock setup into tests/kodi_mocks.py::install_kodi_mocks() so it
can be shared by both tests/ and the new tests-extensive/ conftest.  Move
tests/extreme/ wholesale and the five marked test files to tests-extensive/,
update all tests/extreme path references in the two remaining tests/ files,
and route the four extensive justfile recipes through {{uvdev}} python -m pytest
tests-extensive/.  The extreme harness conftest is renamed to _fixtures.py to
avoid pytest 9's double-registration error when the same conftest.py would be
loaded both hierarchically and via pytest_plugins.
