+++ tests-extensive/extreme/fixtures/addon-settings-template.xml
