+++ tests-extensive/extreme/fixtures/guisettings.xml
