+++ tests-extensive/extreme/scripts/_storage_discovery.py
