+++ tests-extensive/extreme/scripts/install_jurialmunkey_repo.sh
